Modularize Julia
- Reduce compilation time
