+++ src/julia/Prechelt/Project.toml
@@ -1,0 +1,3 @@
+name = "Prechelt"
+uuid = "25fd9c94-8b2a-41fe-8464-d6ac6e1306e3"
+version = "0.1.0"
